conf.d/forgit.plugin.fish: Remove emojify support
The 'eval "$cmd"' in forgit::log takes forever to complete in large
repositories because the command has to finish before the output appears
in fzf. The 'eval' only happens because of the emojify support, which I
am never going to use so just remove support for it altogether to speed
up forgit::log.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>